Phase 2 / S15.a: handoff CLI hotspot decomposition (1119 -> 798 LOC)
Extract closure-captured helper bundle and the --status / --abort
branches from cli/src/commands/handoff.ts into a new
cli/src/commands/handoff/helpers.ts module (factory pattern).

handoff.ts: 1119 -> 798 LOC, under the §15 hotspot cap.

No behavioural change. The closure captures (containerName, targetNs,
aksPfPort, aksPfProc) migrate from action-scope to factory-scope; the
helper functions still see the same set of variables, the same way.

Audit: docs/security-audits/2026-04-29-phase2-hotspot-handoff-cli.md
Tests: 454 passing | 2 skipped. Build + tsc + lint all clean.
